fix #1513 - Wrong portal_url used for TinyMCE in multilingual site.
Also some refactoring together with plone.app.layout
